+++ index.js
@@ -5,8 +5,8 @@
 import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
-import NovaConta from './src/screens/NovaConta';
-import RecuperarSenha from './src/screens/RecuperarSenha';
+import NovaConta from './src/pages/novaconta';
+import RecuperarSenha from './src/pages/recuperarsenha';
 
 
 AppRegistry.registerComponent(appName, () => App);
