+++ src/pages/novaconta/index.js
@@ -1,7 +1,8 @@
 import { useState } from 'react';
 import { StyleSheet, View } from 'react-native';
-import { DefaultButton } from '../components/DefaultButton';
-import DefaultInput from '../components/DefaultInput';
+import { DefaultButton } from '../../components/DefaultButton';
+import DefaultInput from '../../components/DefaultInput';
+import styles from './styles';
 
 export default function NovaConta() {
   const [email, setEmail] = useState('')
@@ -38,28 +39,3 @@ export default function NovaConta() {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-
-  container: {
-    backgroundColor: '#372775',
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 20
-  },
-
-  input: {
-    borderWidth: 1,
-    borderRadius: 20,
-    padding: '2%',
-    width: 300,
-    color: 'white',
-    borderColor: 'white'
-  },
-
-  bottomContainer:{
-    gap: 10,
-    marginTop: 50
-  }
-});
