+++ src/pages/novapesquisa/index.js
@@ -1,8 +1,9 @@
 import React, { useState } from 'react';
 import { StyleSheet, View, Text } from 'react-native';
-import { DefaultButton } from '../components/DefaultButton';
-import DefaultInput from '../components/DefaultInput';
+import { DefaultButton } from '../../components/DefaultButton';
+import DefaultInput from '../../components/DefaultInput';
 import { DefaultSectionT } from 'react-native';
+import styles from './styles';
 
 
 
@@ -25,34 +26,3 @@ export default function NovaPesquisa(props) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  forms: {
-    paddingTop: 30,
-  },
-  busca: {
-    paddingLeft: 100,
-  },
-  container: {
-    backgroundColor: '#372775',
-    flex: 1,
-    alignItems: 'center', 
-    justifyContent: 'flex-start',
-    gap: 20, 
-  },
-  botao: {
-    flex: 1, 
-    justifyContent: 'flex-end', 
-    marginBottom: 20, 
-  },
-  content1: {
-    color:'black'
-  },
-  errorText: {
-    color: 'red',
-    fontSize: 12,
-    marginTop: 1,
-    marginBottom: 10,
-  },
-
-});
